pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dp-fedmed"
version = "0.2.0"
description = "Differential Privacy + Federated Learning for Medical Imaging with SSL Pretraining (Flower + Opacus + MONAI + LightlySSL)"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]

dependencies = [
    "flwr[simulation]>=1.11.0",
    "opacus>=1.5.0",
    "monai>=1.3.0",
    "torch>=2.0.0",
    "torchvision",
    "numpy",
    "pillow",
    "tomli",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "lightly>=1.5.22",
]

[project.optional-dependencies]
benchmark = [
    "matplotlib",
    "pandas",
]
dev = [
    "pytest",
    "black",
    "ruff",
    "pre-commit",
    "ty",
    "basedpyright",
    "bandit[toml]",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["dp_fedmed"]

# ============================================================================
# DEFAULT APP: Downstream Federated Learning (use: flwr run .)
# ============================================================================
[tool.flwr.app]
publisher = "dp-fedmed"

[tool.flwr.app.components]
serverapp = "dp_fedmed.fl.server.app:app"
clientapp = "dp_fedmed.fl.client.app:app"

[tool.flwr.app.config]
config-file = "configs/default.toml"

# ============================================================================
# SSL PRETRAINING: Use the script instead
#   python scripts/run_ssl_pretraining.py configs/pretraining.toml
# ============================================================================

# ============================================================================
# FEDERATION CONFIGURATIONS
# ============================================================================
[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 2
options.backend.client-resources.num-cpus = 1
options.backend.client-resources.num-gpus = 0.3
options.backend.init-args.include_dashboard = false

[tool.flwr.federations.ssl-simulation]
options.num-supernodes = 2
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0.3
options.backend.init-args.include_dashboard = false

[tool.bandit]
exclude_dirs = [".venv", "tests", "__pycache__"]
skips = ["B101"]  # Skip assert warnings (common in tests)

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
reportMissingImports = false
reportMissingTypeStubs = false
reportUnsupportedDunderAll = false
reportOptionalCall = false
